refactor(client): isolate notifications and selection
